Move `generateMdFileEntry`/`updateReadme` functions to separated module
