+++ bin/docgen/index.ts
@@ -2,13 +2,13 @@ import * as A from "fp-ts/lib/Array.js"
 import { pipe } from "fp-ts/lib/function.js"
 import * as Ord from "fp-ts/lib/Ord.js"
 import * as string from "fp-ts/lib/string.js"
-import { readFile, writeFile } from "fs/promises"
 import glob from "glob"
 import path from "path"
 
 import { FileProperties } from "../../types"
 
 import { parseComments } from "./parseComments.js"
+import { generateMdFileEntry, updateReadme } from "./readmeMarkdown.js"
 
 const getFiles = (): string[] => {
   try {
@@ -17,34 +17,6 @@ const getFiles = (): string[] => {
     console.error(err)
     return []
   }
-}
-
-const generateMdFileEntry = ({
-  filename,
-  title,
-  description,
-}: FileProperties): string => {
-  return `
-### [${title}](https://raw.githubusercontent.com/mkobayashime/bookmarklets/main/dist/${filename.replace(
-    /.ts$/,
-    ".js"
-  )})
-
-${description ?? ""}
-  `.trim()
-}
-
-const updateReadme = async (scriptsMarkdown: string): Promise<void> => {
-  const readme = (await readFile(path.resolve(".", "README.md"))).toString()
-  if (!readme) return
-
-  const readmeCommonPart = readme.slice(
-    0,
-    readme.indexOf("## Scripts") + "## Scripts".length
-  )
-
-  const updatedReadme = readmeCommonPart + "\n\n" + scriptsMarkdown
-  await writeFile(path.resolve("README.md"), updatedReadme)
 }
 
 //
